+++ R/connection_spark.R
@@ -112,6 +112,9 @@ spark_connect <- function(master = "local",
   if (master == "spark://HOST:PORT")
     method <- "test"
 
+  if (spark_master_is_gateway(master))
+    method <- "gateway"
+
   # spark-shell (local install of spark)
   if (method == "shell") {
     scon <- shell_connection(master = master,
@@ -121,7 +124,16 @@ spark_connect <- function(master = "local",
                              hadoop_version = hadoop_version,
                              shell_args = shell_args,
                              config = config,
-                             service = FALSE,
+                             service = spark_config_value(
+                               config,
+                               "sparklyr.gateway.service",
+                               FALSE),
+                             remote = spark_config_value(
+                               config,
+                               "sparklyr.gateway.remote",
+                               # running in yarn-cluster mode requires the
+                               # driver to take external connections
+                               TRUE),
                              extensions = extensions)
   } else if (method == "livy") {
     scon <- livy_connection(master = master,
@@ -130,6 +142,8 @@ spark_connect <- function(master = "local",
                             version,
                             hadoop_version ,
                             extensions)
+  } else if (method == "gateway") {
+    scon <- gateway_connection(master = master, config = config)
   } else if (method == "databricks") {
     scon <- databricks_connection(config = config,
                                   extensions)
@@ -292,6 +306,14 @@ spark_connection_is_yarn_client <- function(sc) {
 
 spark_master_is_yarn_client <- function(master) {
   grepl("^yarn-client$", master, ignore.case = TRUE, perl = TRUE)
+}
+
+spark_master_is_yarn_cluster <- function(master) {
+  grepl("^yarn-cluster$", master, ignore.case = TRUE, perl = TRUE)
+}
+
+spark_master_is_gateway <- function(master) {
+  grepl("sparklyr://.*", master)
 }
 
 # Number of cores available in the local install